requirements.txt
# ── Core ───────────────────────────────────────────────────
Django==4.2.16
gunicorn==21.2.0

# ── ASGI server (WebSocket / chat support) ─────────────────
daphne==4.1.2
channels==4.0.0
channels-redis==4.2.0

# ── Database ───────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ── Auth ───────────────────────────────────────────────────
django-allauth==0.57.2

# ── Config / environment ───────────────────────────────────
python-decouple==3.8

# ── Static files ───────────────────────────────────────────
whitenoise==6.6.0

# ── PDF generation ─────────────────────────────────────────
reportlab==4.2.5

# ── Redis client ───────────────────────────────────────────
redis==5.0.1

# ── Image handling (10.4.0 supports Python 3.13+) ──────────
Pillow==10.4.0
